+++ app/Core/XHRProvider.ts
@@ -21,7 +21,7 @@ export class XHRProvider extends HttprProvider {
     return new Promise((resolve, reject) => {
       let xhr = new XMLHttpRequest(),
         queryParams = urlEncode(settings.params),
-        url = queryParams ? urlJoin(settings.url, queryParams) : settings.url;
+        url = queryParams ? urlJoin(settings.url, `?${queryParams}`) : settings.url;
 
       if (_.startsWith(url, '/')) {
         url = `${location.origin}${url}`;
@@ -42,6 +42,11 @@ export class XHRProvider extends HttprProvider {
           headers = _.fromPairs(xhr.getAllResponseHeaders()
           .split('\n')
           .map((header) => header.split(':').map((part) => part.trim()))
+          .map((data) => {
+            data[0] = data[0].toLowerCase();
+
+            return data;
+          })
           .filter((header) => !!header[0])) as StringMap;
 
           response = {